+++ .eslintrc.json
@@ -38,6 +38,7 @@
       2,
       "never"
     ],
+    "space-in-brackets": [2, "never"],
     "array-callback-return": 2,
     "arrow-body-style": 0,
     "arrow-parens": [
