+++ src/app/app.ts
@@ -50,74 +50,6 @@ class App {
         };
         return ticket;
     }
-
-    // getCreateFormResponse returns a form response to create a ticket from a post
-    getCreateFormResponse(message: string): AppCallResponse {
-        const response: AppCallResponse = {
-            type: 'form',
-            form: {
-                title: 'Create Zendesk Ticket',
-                header: 'Create a Zendesk from Mattermost by filling in the required fields and clicking the submit button. Addition text can be added in the `Optional message` form.',
-                footer: 'Message modal form footer',
-                fields: [
-                    {
-                        name: 'subject',
-                        description: 'zendesk subject',
-                        type: 'text',
-                        is_required: true,
-                        label: 'User',
-                        hint: 'Zendesk ticket subject',
-                        position: 1,
-                        modal_label: 'Ticket subject',
-                    },
-                    {
-                        name: 'type',
-                        type: 'static_select',
-                        options: [
-                            {
-                                label: 'problem',
-                                value: 'problem',
-                            },
-                            {
-                                label: 'incident',
-                                value: 'incident',
-                            },
-                        ],
-                        is_required: true,
-                        label: 'User',
-                        hint: 'Zendesk ticket type',
-                        position: 1,
-                        modal_label: 'Ticket type',
-                    },
-                    {
-                        name: 'additional_message',
-                        description: 'zendesk additional message',
-                        type: 'text',
-                        label: 'message',
-                        hint: 'Add additional message to the Zendesk ticket',
-                        modal_label: 'Optional message',
-                        subtype: 'textarea',
-                        min_length: 2,
-                        max_length: 1024,
-                    },
-                    {
-                        name: 'post_message',
-                        description: 'zendesk',
-                        type: 'text',
-                        is_required: true,
-                        value: message,
-                        label: 'message',
-                        modal_label: 'Mattermost message',
-                        subtype: 'textarea',
-                        min_length: 2,
-                        max_length: 1024,
-                    },
-                ],
-
-            },
-        };
-        return response;
-    }
 }
 
 export default new App();
