+++ src/restapi/zd_routes.ts
@@ -1,4 +1,4 @@
-import express, { Request, Response } from 'express';
+import express, {Request, Response} from 'express';
 
 import app from '../app/app';
